+++ CHANGES.md
@@ -1,3 +1,8 @@
+# UNRELEASED
+
+- Trace:
+  - Add information about added clauses to Implication cards
+
 # v3.7.2 (July 2026)
 
 Requires Menhir 20211230 and OCaml 4.14 or above on Linux, Windows and